Extract BloodGlucosePresenter
