+++ app/src/main/java/im/rah/nightwear/NightWearComplicationProviderService.kt
@@ -39,7 +39,7 @@ class NightWearComplicationProviderService : ComplicationProviderService() {
         val builder = ComplicationData.Builder(type)
         
         if (type == ComplicationData.TYPE_SHORT_TEXT) {
-            val bgText: String = bg.combinedString(mmol = mmol, markOld = true, saferUnicode = true)
+            val bgText: String = BloodGlucosePresenter(bg).combinedString(mmol = mmol, markOld = true, saferUnicode = true)
 
             Log.d(TAG, "updating complication data (SHORT_TEXT), bgText: " + bgText)
 
@@ -58,7 +58,7 @@ class NightWearComplicationProviderService : ComplicationProviderService() {
         }
 
         if (type == ComplicationData.TYPE_LONG_TEXT) {
-            val bgText: String = bg.combinedString(mmol = mmol, markOld = false, saferUnicode = true) + " (^1)" // the ^1 is interpolated with the age
+            val bgText: String = BloodGlucosePresenter(bg).combinedString(mmol = mmol, markOld = false, saferUnicode = true) + " (^1)" // the ^1 is interpolated with the age
 
             Log.d(TAG, "updating complication data (LONG_TEXT), bgText: " + bgText)
